testes para 8 funcoes do backend
