+++ busca-keywords/dados/BOM REPOUSO.json
@@ -13,5 +13,155 @@
         "Data": "20-Outubro-2023",
         "Categoria": "Crédito Suplementar",
         "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
+    },
+    {
+        "Data": "20-Outubro-2023",
+        "Categoria": "Crédito Suplementar",
+        "Soma": 30000.0
     }
 ]
