+++ busca-keywords/tests/dados_test/teste_arquivo.json
@@ -1,6 +1,6 @@
 [
     {
-        "Data": "01-January-2023",
+        "Data": "01-Janeiro-2023",
         "Categoria": "Teste",
         "Soma": 15800.0
     }
