+++ busca-keywords/tests/test_create_json.py
@@ -8,7 +8,7 @@ def test_create_json():
     # Testa se a função create_json cria corretamente um arquivo JSON
     nome_arquivo = 'teste_arquivo'
     pasta = 'busca-keywords/tests/dados_test'
-    dados = [{'Data': '01-January-2023', 'Categoria': 'Teste', 'Soma': 15800.0}]
+    dados = [{'Data': '01-Janeiro-2023', 'Categoria': 'Teste', 'Soma': 15800.0}]
     create_json(nome_arquivo, dados, pasta)
 
     # Verifica se o arquivo JSON foi criado
