+++ busca-keywords/tests/test_extracao_trechos/test_full/test_file.txt
@@ -1,4 +1,4 @@
-PREFEITURA DE MunicipioX
+PREFEITURA DE MUNICIPIOX
 12 de Janeiro de 2023
 
 Texto com a palavra-chave CRÉDITO SUPLEMENTAR no valor de R$ 1.994,70 .
