+++ busca-keywords/tests/test_extracao_trechos/test_trechos/MUNICIPIOX-test_file_trechos.txt
@@ -1,0 +1,16 @@
+12 de Janeiro de 2023
+MUNICIPIOX
+have CRÉDITO SUPLEMENTAR no valor de R$ 1.994,70 .
+
+Texto com a palavra-chave CRÉDITO SUPLEMENTAR no valor de R$ 35.000,10 .
+12 de Janeiro de 2023
+MUNICIPIOX
+have CRÉDITO SUPLEMENTAR no valor de R$ 35.000,10 .
+12 de Janeiro de 2023
+MUNICIPIOX
+have CRÉDITO SUPLEMENTAR no valor de R$ 1.994,70 .
+
+Texto com a palavra-chave CRÉDITO SUPLEMENTAR no valor de R$ 35.000,10 .
+12 de Janeiro de 2023
+MUNICIPIOX
+have CRÉDITO SUPLEMENTAR no valor de R$ 35.000,10 .
