+++ busca-keywords/tests/test_extracao_trechos/test_trechos/test_file_trechos.txt
@@ -1,0 +1,6 @@
+12 de Janeiro de 2023
+MUNICIPIOX
+Trecho 1
+12 de Janeiro de 2023
+MUNICIPIOX
+Trecho 2
