+++ busca-keywords/tests/trechos_test/MUNICIPIOX-20-Outubro-2023-municipios-mb.txt
@@ -1,3 +1,3 @@
 20 de Outubro de 2023
-MunicipioX
+MUNICIPIOX
 Texto da data de contendo o valor de R$ 1.000,00
